+++ bentley_ottmann/hints.py
@@ -1,5 +1,8 @@
 from decimal import Decimal
 from numbers import Real
-from typing import TypeVar
+from typing import (Tuple,
+                    TypeVar)
 
 Scalar = TypeVar('Scalar', Real, Decimal)
+Point = Tuple[Scalar, Scalar]
+Segment = Tuple[Point, Point]
