+++ tests/core_tests/linear_tests/test_find_intersections.py
@@ -2,8 +2,8 @@ from typing import Tuple
 
 from hypothesis import given
 
-from bentley_ottmann.linear import (Segment,
-                                    find_intersections)
+from bentley_ottmann.core.linear import (find_intersections)
+from bentley_ottmann.hints import Segment
 from tests.utils import (is_point,
                          reverse_segment)
 from . import strategies
