+++ tests/planar_tests/strategies.py
