+++ tests/planar_tests/test_segments_intersect.py
@@ -2,9 +2,9 @@ from typing import List
 
 from hypothesis import given
 
-from bentley_ottmann.base import segments_intersect
-from bentley_ottmann.linear import (Segment,
-                                    find_intersections)
+from bentley_ottmann.core.linear import (find_intersections)
+from bentley_ottmann.hints import Segment
+from bentley_ottmann.planar import segments_intersect
 from tests.utils import (reverse_segment,
                          reverse_segment_coordinates)
 from . import strategies
