+++ tests/planar_tests/test_segments_intersections.py
@@ -3,9 +3,9 @@ from typing import List
 
 from hypothesis import given
 
-from bentley_ottmann.base import segments_intersections
-from bentley_ottmann.linear import (Segment,
-                                    find_intersections)
+from bentley_ottmann.core.linear import (Segment,
+                                         find_intersections)
+from bentley_ottmann.planar import segments_intersections
 from tests.utils import is_point
 from . import strategies
